# Remove the `Either Int` from value2term