@inariksit 2020-11-29
@johnjcamilleri In the old version, where `value2term` returned an `Either Int Term`, the potential error ("variable #n is out of scope") was caught in 5 places with almost identical behaviour: [1](https://github.com/GrammaticalFramework/gf-core/pull/87/files#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fL36), [2](https://github.com/GrammaticalFramework/gf-core/pull/87/files#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fL291-L292), [3](https://github.com/GrammaticalFramework/gf-core/pull/87/files#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fL358-L359), [4](https://github.com/GrammaticalFramework/gf-core/pull/87/files#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fL358-L359), [5](https://github.com/GrammaticalFramework/gf-core/pull/87/files#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fL495-L496). I've never seen that error in the wild, but maybe @aarneranta or @krangelov can explain when it's supposed to be triggered, so we can test.

It looks to me that the PGFs are identical. I ran gftest on an application grammar as a part of my normal development process (store old PGF compiled with old GF -> implement new feature -> compile with new GF -> gftest the two versions) and got the expected results, that is, only intended changes. I also linearised manually some ad hoc test sentences on the PGF compiled with the new GF, and was happy with them.
Here's a more rigorous test with FoodsEng:

```
➜ gf -make FoodsEng.gf
Writing Foods.pgf...
➜ mv Foods.pgf /tmp   

➜ gf-old-slow -make FoodsEng.gf
Writing Foods.pgf...

➜ gftest -g Foods.pgf -o /tmp/Foods.pgf 
<TL;DR: identical result>

➜ md5sum Foods.pgf
0e5e30e31e0976cae23884fad2c3b784  Foods.pgf
➜ md5sum /tmp/Foods.pgf
0e5e30e31e0976cae23884fad2c3b784  /tmp/Foods.pgf
```