@anka-213 2020-12-02
This change should not change the behaviour in any way other than with respect to performance (and possibly error messages), since all I've done is remove the `Either Int` wrapping from the function `value2term`, replaced the `Left` case with a call to `error` and replaced all the monadic functions in it with their non-monadic counterparts.

The only ways it could change anything is if some code relies on the error of the `Left` case being thrown swiftly instead of eventually (or possibly never, if the `Term` is not fully evaluated). My guess is that the `Left` case would correspond to a compiler bug and not a user error, but I may be mistaken. If it was user-facing, the UX of `"variable #637 is out of scope"` would be pretty bad.