@johnjcamilleri 2020-12-02
Sounds reasonable, Andreas. I would make the same assumption about it indicating a compiler bug.

Since the performance gains are so huge, we should really push this through for inclusion in the next release. But before merging, it would be nice to get the blessing of someone who knows the internals well. Looking at the git history, seems it was @krangelov who introduced the `Either Int` to `value2term` [here](https://github.com/GrammaticalFramework/gf-core/commit/07cf4d6509#diff-f9129837a82356260e00c61f38abfed49b25b2a5ae85b47ba17741add187885fR489).