@johnjcamilleri 2020-12-10
I finally got around to testing the version of GF, however I'm disappointed to say that I haven't noticed any improvement at all 😞 
I tried two things:
1. building the RGL to GFOs
2. compiling a large application grammar (to PGF) which normally takes 90s and 5.56 GB

In both cases I saw identical time and space usage with the old and new GF versions. I'm pretty sure I was _not_ using the same GF executable for both, as I changed the version number in `gf.cabal` for this version and confirmed with `gf --version` before each test.

So, any idea why in certain cases this gives no improvement at all? Or maybe I'm doing something wrong?
Can someone else run a test and produce a noticeable improvement?